# openapi-spec.yaml
openapi: 3.0.3
info:
  title: SIRAJ v6.1 Computational Hermeneutics MCP Server API
  version: 6.1.0
  description: |
    ## Overview
    
    SIRAJ v6.1 Computational Hermeneutics MCP Server provides advanced linguistic analysis capabilities through the Model Context Protocol (MCP). This API enables deep semantic understanding of Arabic and Semitic languages with cultural sovereignty protection.
    
    ## Protocol
    
    This server implements the Model Context Protocol (MCP) specification using JSON-RPC 2.0 messages. The actual communication occurs over stdio or Server-Sent Events (SSE) transports, not traditional HTTP REST endpoints.
    
    ## Key Features
    
    - **72-Node Archetypal Semantic Mapping**: Deep semantic understanding through archetypal analysis
    - **Cultural Sovereignty Protection**: Community-driven validation for sensitive content
    - **Multi-Paradigm Validation**: Traditional, scientific, and computational approaches
    - **Transformer Integration**: Support for multilingual BERT and Arabic-specific models
    - **Real-time Processing**: Efficient caching and batch processing capabilities
    
    ## MCP Components
    
    This server provides:
    - **Tools**: Executable functions for linguistic analysis
    - **Resources**: Access to linguistic databases and corpora
    - **Prompts**: Pre-built templates for common analysis workflows
    
    ## Transport Protocols
    
    - **stdio**: Default transport for Claude Desktop integration
...
servers:
...
paths:
  /initialize:
    post:
      tags: [initialization]
      summary: Initialize MCP connection
      description: |
        Initialize the MCP connection and negotiate capabilities. This is the first message sent by the client to establish the connection.
      operationId: initialize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitializeRequest'
            example:
              jsonrpc: "2.0"
              id: 1
              method: "initialize"
              params:
                protocolVersion: "2024-11-05"
                capabilities:
                  experimental: {}
                  sampling: {}
                clientInfo:
                  name: "Claude Desktop"
                  version: "1.0.0"
      responses:
        '200':
          description: Initialization successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitializeResponse'
              example:
                jsonrpc: "2.0"
                id: 1
                result:
                  protocolVersion: "2024-11-05"
                  capabilities:
                    tools: {}
                    resources:
                      subscribe: true
                      listChanged: true
                    prompts:
                      listChanged: true
                    logging: {}
                  serverInfo:
                    name: "siraj-computational-hermeneutics"
                    version: "6.1.0"
                    description: "SIRAJ v6.1 Computational Hermeneutics MCP Server"
...
components:
  schemas:
...
    JSONRPCRequest:
      type: object
      required: [jsonrpc, id, method]
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        id:
          oneOf:
            - type: string
            - type: number
        method:
          type: string
        params:
          type: object
    JSONRPCResponse:
      type: object
      required: [jsonrpc, id]
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        id:
          oneOf:
            - type: string
            - type: number
        result:
          type: object
        error:
          $ref: '#/components/schemas/JSONRPCError'
    JSONRPCError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object

    # MCP Core schemas
    ClientCapabilities:
      type: object
      properties:
        experimental:
          type: object
        sampling:
          type: object
    ServerCapabilities:
      type: object
      properties:
        tools:
          type: object
          properties:
            listChanged:
              type: boolean
        resources:
          type: object
          properties:
            subscribe:
              type: boolean
            listChanged:
              type: boolean
        prompts:
          type: object
          properties:
            listChanged:
              type: boolean
        logging:
          type: object
    ClientInfo:
      type: object
      required: [name, version]
      properties:
        name:
          type: string
        version:
          type: string
    ServerInfo:
      type: object
      required: [name, version]
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string

    # Tool schemas
...
    InitializeRequest:
      allOf:
        - $ref: '#/components/schemas/JSONRPCRequest'
        - type: object
          properties:
            method:
              enum: ["initialize"]
            params:
              type: object
              required: [protocolVersion, capabilities, clientInfo]
              properties:
                protocolVersion:
                  type: string
                capabilities:
                  $ref: '#/components/schemas/ClientCapabilities'
                clientInfo:
                  $ref: '#/components/schemas/ClientInfo'
    InitializeResponse:
      allOf:
        - $ref: '#/components/schemas/JSONRPCResponse'
        - type: object
          properties:
            result:
              type: object
              required: [protocolVersion, capabilities, serverInfo]
              properties:
                protocolVersion:
                  type: string
                capabilities:
                  $ref: '#/components/schemas/ServerCapabilities'
                serverInfo:
                  $ref: '#/components/schemas/ServerInfo'
...